@@ -1,9 +1,9 @@
 package com.assignment.trainenquiry.adapter
 
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
 import com.assignment.kotlinmvvm.DataModel.Android
 import com.assignment.kotlinmvvm.MainActivity
 import com.assignment.kotlinmvvm.R
